io.jl

Public functions

QuantumACES.aces_data_filename — Method

aces_data_filename(aces_data::ACESData)
aces_data_filename(d::Design, budget_set::Vector{Int})
aces_data_filename(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Int, repeat_numbers::Vector{Int}, full_covariance::Bool, ls_type::Symbol, budget_set::Vector{Int})

Returns a string describing the filename corresponding to the ACES data.

QuantumACES.delete_aces — Method

delete_aces(aces_data::ACESData)

Deletes the file corresponding to the ACES data aces_data.

QuantumACES.delete_design — Method

delete_design(d::Design)

Deletes the file corresponding to the design d.

QuantumACES.delete_scaling — Method

delete_scaling(scaling_data::AbstractScalingData)

Deletes the file corresponding to the scaling data scaling_data.

QuantumACES.design_filename — Method

design_filename(d::Design)
design_filename(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Int, repeat_numbers::Vector{Int}, full_covariance::Bool, ls_type::Symbol)

Returns a string describing the filename corresponding to the supplied design data.

QuantumACES.enter_folder — Method

enter_folder(folder::String)

If not currently in the folder folder, enter it.

QuantumACES.exit_folder — Method

exit_folder(folder::String)

If currently in the folder folder, exit it.

QuantumACES.load_aces — Method

load_aces(d::Design, budget_set::Vector{Int})
load_aces(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Int, repeat_numbers::Vector{Int}, full_covariance::Bool, ls_type::Symbol, budget_set::Vector{Int})

Loads the ACES data whose filename is specified by the supplied variables.

QuantumACES.load_design — Method

load_design(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Int, repeat_numbers::Vector{Int}, full_covariance::Bool, ls_type::Symbol)

Loads the design whose filename is specified by the supplied variables.

QuantumACES.load_scaling — Method

load_scaling(d::Design, ls_type::Symbol)
load_scaling(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Int, repeat_numbers::Vector{Int}, ls_type::Symbol)

Loads the scaling data whose filename is specified by the supplied variables.

QuantumACES.save_aces — Method

save_aces(aces_data::ACESData; clear_design::Bool = false)

Saves the ACES data aces_data with the appropriate filename, and deletes the design file if it exists and clear_design is true.

QuantumACES.save_design — Method

save_design(d::Design)

Saves the design d with the appropriate filename.

QuantumACES.save_scaling — Method

save_scaling(scaling_data::AbstractScalingData)

Saves the scaling data scaling_data with the appropriate filename.

QuantumACES.scaling_filename — Method

scaling_filename(scaling_data::AbstractScalingData)
scaling_filename(d::Design, ls_type::Symbol)
scaling_filename(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Int, repeat_numbers::Vector{Int}, ls_type::Symbol)

Returns a string describing the filename for the scaling data corresponding to the supplied design data.

Private functions

QuantumACES.tuples_filename — Method

tuples_filename(tuple_number::Int, repeat_numbers::Vector{Int})

Returns a string describing the filename for the supplied tuple set data.