io.jl

Public functions

QuantumACES.aces_data_filename — Method

aces_data_filename(aces_data::ACESData)
aces_data_filename(d::Design, budget_set::Vector{Int})
aces_data_filename(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Integer, repeat_numbers::Vector{Int}, full_covariance::Bool, ls_type::Symbol, budget_set::Vector{Int})

Returns a string describing the filename corresponding to the ACES data.

QuantumACES.delete_aces — Method

delete_aces(aces_data::ACESData)

Deletes the file corresponding to the ACES data aces_data.

QuantumACES.delete_design — Method

delete_design(d::Design)

Deletes the file corresponding to the design d.

QuantumACES.delete_qiskit_ensemble — Method

delete_qiskit_ensemble(d_rand::RandDesign)

Deletes the Qiskit ensemble whose filename is specified by the supplied randomised experimental design.

QuantumACES.delete_rand_design — Method

delete_rand_design(d_rand::RandDesign)

Deletes the file corresponding to the randomised experimental design d_rand.

QuantumACES.delete_rand_design_job — Method

delete_rand_design_job(d_rand::RandDesign, backend::String, job_idx::Integer)

Deletes the job counts data for the randomised experimental design d_rand whose filename is specified by the supplied variables.

QuantumACES.delete_scaling — Method

delete_scaling(scaling_data::AbstractScalingData)

Deletes the file corresponding to the scaling data scaling_data.

QuantumACES.design_filename — Method

design_filename(d::Design)
design_filename(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Integer, repeat_numbers::Vector{Int}, full_covariance::Bool, ls_type::Symbol)

Returns a string describing the filename corresponding to the supplied design data.

QuantumACES.enter_folder — Method

enter_folder(folder::String)

If not currently in the folder folder, enter it.

QuantumACES.exit_folder — Method

exit_folder(folder::String)

If currently in the folder folder, exit it.

QuantumACES.load_aces — Method

load_aces(d::Design, budget_set::Vector{Int})
load_aces(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Integer, repeat_numbers::Vector{Int}, full_covariance::Bool, ls_type::Symbol, budget_set::Vector{Int})

Loads the ACES data whose filename is specified by the supplied variables.

QuantumACES.load_design — Method

load_design(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Integer, repeat_numbers::Vector{Int}, full_covariance::Bool, ls_type::Symbol)

Loads the design whose filename is specified by the supplied variables.

QuantumACES.load_qiskit_ensemble — Method

load_qiskit_ensemble(d_rand::RandDesign)

Loads the Qiskit ensemble whose filename is specified by the supplied randomised experimental design.

QuantumACES.load_rand_design — Method

load_rand_design(d::Design, total_randomisations::Integer, seed::UInt64)
load_rand_design(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Integer, repeat_numbers::Vector{Int}, full_covariance::Bool, ls_type::Symbol, total_randomisations::Integer, seed::UInt64)

Loads the randomised experimental design whose filename is specified by the supplied variables.

QuantumACES.load_rand_design_job — Method

load_rand_design_job(d_rand::RandDesign, backend::String, job_idx::Integer)

Loads the job counts data for the randomised experimental design d_rand whose filename is specified by the supplied variables.

QuantumACES.load_scaling — Method

load_scaling(d::Design, ls_type::Symbol)
load_scaling(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Integer, repeat_numbers::Vector{Int}, ls_type::Symbol)

Loads the scaling data whose filename is specified by the supplied variables.

QuantumACES.rand_design_filename — Method

rand_design_filename(d_rand::RandDesign)
rand_design_filename(d::Design, total_randomisations::Integer, seed::UInt64)
rand_design_filename(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Integer, repeat_numbers::Vector{Int}, full_covariance::Bool, ls_type::Symbol, total_randomisations::Integer, seed::UInt64)

Returns a string describing the filename corresponding to the randomised design data.

QuantumACES.save_aces — Method

save_aces(aces_data::ACESData; clear_design::Bool = false)

Saves the ACES data aces_data with the appropriate filename, and deletes the design file if it exists and clear_design is true.

QuantumACES.save_design — Method

save_design(d::Design)

Saves the design d with the appropriate filename.

QuantumACES.save_qiskit_ensemble — Method

save_qiskit_ensemble(d_rand::RandDesign, qiskit_ensemble::Py)

Saves the Qiskit ensemble qiskit_ensemble with the appropriate filename.

QuantumACES.save_rand_design — Method

save_rand_design(d_rand::RandDesign)

Saves the randomised experimental design d_rand with the appropriate filename.

QuantumACES.save_rand_design_job — Method

save_rand_design_job(d_rand::RandDesign, backend::String, job_counts::Union{Vector{Matrix{UInt8}}, Vector{Dict{String, Int}}}, job_idx::Integer)

Saves the job counts data job_counts with index job_idx for the randomised experimental design d_rand with the filename determined by the supplied backend.

QuantumACES.save_scaling — Method

save_scaling(scaling_data::AbstractScalingData)

Saves the scaling data scaling_data with the appropriate filename.

QuantumACES.scaling_filename — Method

scaling_filename(scaling_data::AbstractScalingData)
scaling_filename(d::Design, ls_type::Symbol)
scaling_filename(circuit_param::AbstractCircuitParameters, noise_param::AbstractNoiseParameters, tuple_number::Integer, repeat_numbers::Vector{Int}, ls_type::Symbol)

Returns a string describing the filename for the scaling data corresponding to the supplied design data.

Private functions

QuantumACES.tuples_filename — Method

tuples_filename(tuple_number::Integer, repeat_numbers::Vector{Int})

Returns a string describing the filename for the supplied tuple set data.