utils.jl

Public functions

QuantumACES.get_mapping_string — Method

get_mapping_string(m::Mapping, c::AbstractCircuit; two_qubit_only::Bool = false)

Returns the string representation of the mapping m for the circuit c, including eigenvalues.

QuantumACES.get_pauli_string — Method

get_pauli_string(p::Pauli)

Returns the string representation of the Pauli p.

QuantumACES.get_support — Method

get_support(p::Pauli)

Returns the support of the Pauli p.

QuantumACES.pretty_print — Method

pretty_print(aces_data::ACESData, merit_set::Tuple{Merit, Merit, Merit})

Prints the z-scores of the normalised RMS errors of the gate eigenvalue estimator vector for the GLS, WLS, and OLS estimators in aces_data using the predicted means and variances for each in merit_set.

QuantumACES.pretty_print — Method

pretty_print(d::Design)

Prints the tuple set and shot weight data of the design d in a readable format.

QuantumACES.pretty_print — Method

pretty_print(merit_array::Matrix{Float64})

Prints merit_array, produced by compare_ls_optimise_weights, in a readable format.

Private functions

QuantumACES.project_simplex — Method

project_simplex(probabilities::Vector{Float64})

Returns a copy of the probability distribution probabilities projected into the probability simplex according to the Euclidean norm.

QuantumACES.wht_matrix — Method

wht_matrix(n::Int)

Returns the symplectically ordered Walsh-Hadamard transform matrix of order n, which maps an n-qubit Pauli error probability distribution to its eigenvalues. The inverse transform is obtained by dividing the transform by the factor $4^n$.